app att

+++ app.py
@@ -2,12 +2,13 @@ from flask import Flask, request, jsonify
 from models.user import User
 from models.database import db
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
+import bcrypt
 
 app = Flask(__name__)
 # Chave usada para garantir a segurança dos dados da sessão
 app.config["SECRET_KEY"] = "your_secret_key"
 # Define o caminho do banco de dados SQLite
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///database.db"
+app.config['SQLALCHEMY_DATABASE_URI'] = "mysql+pymysql://admin:admin123@127.0.0.1:3306/flask-crud"
 
 login_manager = LoginManager()
 db.init_app(app) # Vincula o banco de dados à instância do app
@@ -28,11 +29,11 @@ def login():
         user = User.query.filter_by(username=username).first() 
         
         # Verifica se o usuário existe e se a senha coincide c a do banco de dados
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             print(current_user.is_authenticated)
             return jsonify({"message": "Autenticado com sucesso!"})
-    return jsonify({"message": "Credenciais inválidas"})
+    return jsonify({"message":"usuario invalido"}), 401
 
 @app.route("/logout", methods=["GET"])
 @login_required
@@ -42,28 +43,20 @@ def logout():
    
 # Começo do Crud (C)
 @app.route("/User", methods=["POST"])
-@login_required
 def create_user():
     data = request.json
     username = data.get("username")
     password = data.get("password")
 
+    if username and password:
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, password=hashed_password)
+        db.session.add(user)
+        db.session.commit()
+        return jsonify({"message":"Usuário criado com sucesso!"})
     
-
-    if not username or not password:
-        return jsonify({"message": "Dados inválidas"}), 400
+    return jsonify({"message":"Credenciais inválidas"}), 400
     
-    user_exist = User.query.filter_by(username=username).first()
-
-    if user_exist:
-        return jsonify({"message": "Esse nome ja está cadastrado"}),400
-    
-     
-    user = User(username=username, password=password)
-    db.session.add(user)
-    db.session.commit()
-    return jsonify({"message": "Usuário cadastrado com sucesso!"}), 201
-
 @app.route("/User/<int:id_user>", methods=["GET"])
 @login_required
 def read_user(id_user):
@@ -80,6 +73,9 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "operação nao permitida"}), 403
+    
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
@@ -94,7 +90,8 @@ def delete_user(id_user):
 
     if id_user == current_user.id:
         return jsonify({"message":"Deleção proibida"}), 403
-    
+    if current_user.role != "admin":
+        return jsonify({"message": "operação nao permitida"})
     if user:
         db.session.delete(user)
         db.session.commit()
